test: lint

+++ contracts/ERC998/interfaces/IERC223Receiver.sol
@@ -10,5 +10,9 @@ interface IERC223Receiver {
    * @param _value Amount of tokens.
    * @param _data  Transaction metadata.
    */
-  function tokenFallback(address _from, uint _value, bytes memory _data) external;
+  function tokenFallback(
+    address _from,
+    uint256 _value,
+    bytes memory _data
+  ) external;
 }
